+++ src/edu/uclouvain/core/nodus/compute/assign/AllOrNothingAssignment.java
@@ -123,7 +123,7 @@ public class AllOrNothingAssignment extends Assignment {
       int threads = assignmentParameters.getThreads();
 
       // Compute costs
-      if (!virtualNet.computeCosts(0, odClass, assignmentParameters.getCostFunctions(), threads)) {
+      if (!virtualNet.computeCosts(0, odClass, threads)) {
         return false;
       }