+++ src/edu/uclouvain/core/nodus/compute/assign/DynamicTimeDependentAssignment.java
@@ -160,8 +160,7 @@ public class DynamicTimeDependentAssignment extends Assignment {
         int threads = assignmentParameters.getThreads();
 
         // Compute costs
-        if (!virtualNet.computeCosts(
-            0, odClass, currentTimeSlice, assignmentParameters.getCostFunctions(), threads)) {
+        if (!virtualNet.computeCosts(0, odClass, currentTimeSlice, threads)) {
           return false;
         }