+++ src/edu/uclouvain/core/nodus/compute/assign/IncrementalAssignment.java
@@ -126,8 +126,7 @@ public class IncrementalAssignment extends Assignment {
         int threads = assignmentParameters.getThreads();
 
         // (re)Compute costs
-        if (!virtualNet.computeCosts(
-            iteration, odClass, assignmentParameters.getCostFunctions(), threads)) {
+        if (!virtualNet.computeCosts(iteration, odClass, threads)) {
           nodusMapPanel.stopProgress();
           return false;
         }